refactor: move en-quire source + tests into packages/en-quire/
Why: step 7 of #58. Completes the monorepo layout by relocating all
markdown/yaml-specific code and en-quire-owned tools into their own
workspace package so en-core only holds reliability primitives and
en-scribe can be added as a sibling in step 8.

What moved (git mv preserves blame):
- src/document/{markdown-parser, yaml-parser, markdown-strategy,
  yaml-strategy, parser} → packages/en-quire/src/parsers/
- src/tools/{read, write, search, status, governance, admin}/** →
  packages/en-quire/src/tools/**
- src/server.ts → packages/en-quire/src/plugin.ts (registerEnQuireTools)
- src/index.ts → packages/en-quire/src/bin.ts (startup + createServer)
- test/helpers/md-ops.ts → packages/en-quire/test/helpers/

Tests split between packages by dependency direction:
- packages/en-core/test/ — shared, config, rbac, git, search/database,
  security/path-traversal (no en-quire imports)
- packages/en-quire/test/ — everything that uses a parser or a tool
  handler (document/*, search/indexer+sync, security/regex-flags+
  exec-tokenise, server/tool-registration)

Fixture directory stays at repo-root test/fixtures/; tests reference
it via ../../../../../test/fixtures/docs from their new depth.

Deleted src/document/serializer.ts — dead code (no callers), md-only.

Other changes:
- Root package.json becomes a private workspace root
  ("en-quire-monorepo"). Root build/test/lint now orchestrate both
  packages. The publishable @nullproof-studio/en-quire moves to v0.2.0
  in packages/en-quire/ with bin: enquire → dist/bin.js.
- vitest.config.ts discovers tests from packages/*/test/ and keeps the
  alias @nullproof-studio/en-core → packages/en-core/src/index.ts so
  the dev loop doesn't need a core rebuild.
- plugin.ts is now pure registration (no McpServer, no ctx). bin.ts
  owns McpServer creation, registry wiring, and transport handling.
- sync-lock-scope.test.ts mock path becomes ../../../src/shared/
  file-utils.js again now that the test lives inside en-core's... wait
  it's actually in en-quire because it depends on markdown-parser.
  Mock path reaches across to ../../../../en-core/src/shared/
  file-utils.js; kept working.
- tool-registration.test.ts updated to import registerEnQuireTools
  from plugin.ts, and its infrastructure-file filter shrinks to just
  write-helpers.ts (context/registry/runtime now live in en-core).

Verification: npm run build + npm run lint + npm test all green
(361/361) under the new two-package layout.

Step 7 of the en-core/en-quire/en-scribe monorepo split (issue #58).
