+++ packages/en-core/test/unit/shared/encoding.test.ts
