+++ packages/en-core/test/unit/shared/word-count.test.ts
