+++ packages/en-quire/package.json
@@ -1,0 +1,52 @@
+{
+  "name": "@nullproof-studio/en-quire",
+  "version": "0.2.0",
+  "description": "Structured document management for agent systems, with governance. An MCP server.",
+  "type": "module",
+  "main": "dist/bin.js",
+  "bin": {
+    "enquire": "dist/bin.js"
+  },
+  "license": "MIT",
+  "author": "Nullproof Studio",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/nullproof-studio/en-quire.git",
+    "directory": "packages/en-quire"
+  },
+  "keywords": [
+    "mcp",
+    "mcp-server",
+    "markdown",
+    "document-management",
+    "agent",
+    "rbac",
+    "git",
+    "governance",
+    "typescript"
+  ],
+  "engines": {
+    "node": ">=22.0.0"
+  },
+  "scripts": {
+    "build": "tsc",
+    "start": "node dist/bin.js",
+    "dev": "tsx src/bin.ts",
+    "clean": "rm -rf dist"
+  },
+  "dependencies": {
+    "@modelcontextprotocol/sdk": "^1.12.1",
+    "@nullproof-studio/en-core": "0.1.5",
+    "remark-frontmatter": "^5.0.0",
+    "remark-gfm": "^4.0.0",
+    "remark-parse": "^11.0.0",
+    "unified": "^11.0.5",
+    "yaml": "^2.7.1",
+    "zod": "^3.24.4"
+  },
+  "devDependencies": {
+    "@types/node": "^22.15.3",
+    "tsx": "^4.19.4",
+    "typescript": "^5.8.3"
+  }
+}
