+++ packages/en-quire/src/tools/search/doc-search.ts
