+++ packages/en-quire/src/tools/write/doc-delete-section.ts
