+++ packages/en-quire/src/tools/write/doc-insert-text.ts
