+++ packages/en-quire/src/tools/write/doc-rename.ts
