+++ packages/en-quire/test/unit/document/fence-validation.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 Nullproof Studio. MIT License — see LICENSE
 import { describe, it, expect } from 'vitest';
 import { parserRegistry } from '@nullproof-studio/en-core';
-import '../../../src/document/markdown-parser.js';
+import '../../../src/parsers/markdown-parser.js';
 
 const parser = parserRegistry.getParser('test.md');
