+++ packages/en-quire/test/unit/document/preamble.test.ts
@@ -1,6 +1,6 @@
 // Copyright (c) 2026 Nullproof Studio. MIT License — see LICENSE
 import { describe, it, expect } from 'vitest';
-import '../../../src/document/markdown-parser.js';
+import '../../../src/parsers/markdown-parser.js';
 import { parserRegistry } from '@nullproof-studio/en-core';
 import { readSection, deleteSection, buildOutline } from '@nullproof-studio/en-core';
 import { replaceSection, appendToSection } from '../../helpers/md-ops.js';
