+++ packages/en-quire/test/unit/document/section-tree.test.ts
@@ -2,11 +2,11 @@
 import { describe, it, expect } from 'vitest';
 import { readFileSync } from 'node:fs';
 import { resolve } from 'node:path';
-import { parseMarkdown } from '../../../src/document/parser.js';
-import { buildSectionTree } from '../../../src/document/markdown-parser.js';
+import { parseMarkdown } from '../../../src/parsers/parser.js';
+import { buildSectionTree } from '../../../src/parsers/markdown-parser.js';
 import { getBreadcrumb, getSectionPath, flattenTree } from '@nullproof-studio/en-core';
 
-const fixturesDir = resolve(import.meta.dirname, '../../fixtures/docs');
+const fixturesDir = resolve(import.meta.dirname, '../../../../../test/fixtures/docs');
 
 function loadFixture(name: string): string {
   return readFileSync(resolve(fixturesDir, name), 'utf-8');
