+++ packages/en-quire/test/unit/document/yaml-parser.test.ts
@@ -1,10 +1,10 @@
 // Copyright (c) 2026 Nullproof Studio. MIT License — see LICENSE
 import { describe, it, expect } from 'vitest';
-import '../../../src/document/yaml-parser.js';
+import '../../../src/parsers/yaml-parser.js';
 import { parserRegistry } from '@nullproof-studio/en-core';
 import { resolveAddress, resolveSingleSection } from '@nullproof-studio/en-core';
 import { readSection, replaceSection, deleteSection, buildOutline, appendToSection } from '@nullproof-studio/en-core';
-import { yamlStrategy } from '../../../src/document/yaml-strategy.js';
+import { yamlStrategy } from '../../../src/parsers/yaml-strategy.js';
 
 const parser = parserRegistry.getParser('test.yaml');
