+++ packages/en-quire/test/unit/search/sync-lock-scope.test.ts
@@ -5,7 +5,7 @@ import { mkdtempSync, rmSync, writeFileSync, mkdirSync } from 'node:fs';
 import { join } from 'node:path';
 import { tmpdir } from 'node:os';
 import { initSearchSchema } from '@nullproof-studio/en-core';
-import '../../../src/document/markdown-parser.js';
+import '../../../src/parsers/markdown-parser.js';
 
 // Track whether readDocument is called while the db is in a transaction.
 // Invariant under test: during index sync, disk I/O + parsing must NOT happen
@@ -17,14 +17,12 @@ import '../../../src/document/markdown-parser.js';
 const readCallsInTransaction: boolean[] = [];
 let currentDb: Database.Database | null = null;
 
-// Mock the internal file-utils module that en-core's search/sync.ts imports
-// via relative path. Mocking the package barrel (@nullproof-studio/en-core)
-// only intercepts external consumers — not sync.ts's own `../shared/file-utils.js`
-// import. This test moves into packages/en-core/ in step 7 where the mock path
-// becomes package-relative again.
-vi.mock('../../../packages/en-core/src/shared/file-utils.js', async () => {
-  const actual = await vi.importActual<typeof import('../../../packages/en-core/src/shared/file-utils.js')>(
-    '../../../packages/en-core/src/shared/file-utils.js',
+// Mock the internal file-utils module that sync.ts imports via relative path.
+// Mocking the package barrel (@nullproof-studio/en-core) only intercepts
+// external consumers — not sync.ts's own `../shared/file-utils.js` import.
+vi.mock('../../../../en-core/src/shared/file-utils.js', async () => {
+  const actual = await vi.importActual<typeof import('../../../../en-core/src/shared/file-utils.js')>(
+    '../../../../en-core/src/shared/file-utils.js',
   );
   return {
     ...actual,
