+++ vitest.config.ts
@@ -6,7 +6,7 @@ export default defineConfig({
   test: {
     globals: true,
     testTimeout: 10000,
-    include: ['test/**/*.test.ts'],
+    include: ['packages/*/test/**/*.test.ts'],
     alias: {
       '@nullproof-studio/en-core': resolve(__dirname, 'packages/en-core/src/index.ts'),
     },
